Extract and rename AutoCompleteViewWindow (mac)
