+++ src/ui/osx/TogglDesktop/Features/Timer/TimerViewController/TimeEditView.swift
@@ -1,0 +1,19 @@
+//
+//  TimeEditView.swift
+//  TogglDesktop
+//
+//  Created by Andrew Nester on 30.10.2020.
+//  Copyright © 2020 Toggl. All rights reserved.
+//
+
+import Cocoa
+
+class TimeEditView: NSView {
+
+    override func draw(_ dirtyRect: NSRect) {
+        super.draw(dirtyRect)
+
+        // Drawing code here.
+    }
+    
+}
